add population of dropdown on homepage
