+++ .gitignore
@@ -3,3 +3,5 @@ node_modules
 npm-debug.log
 .env
 config.json
+.eslintrc.json
+.eslintignore
