+++ public/index.html
@@ -23,8 +23,6 @@
         <span class="bold">What would you like to do?</span>
         <select id="dropdown">
             <option value="default">Select Activity</option>
-            <option value="skiing">Skiing</option>
-            <option value="beach">Beach</option>
         </select>
     </span>
     <br><br>
@@ -58,6 +56,9 @@
     <br>
     <button type="button" class="btn btn-primary">Submit</button>
     </div>
+    <script src="https://code.jquery.com/jquery-3.4.1.min.js"></script>
+    <script src="js/api.js"></script>
+    <script src="js/index.js"></script>
 </body>
 
 </html>
