+++ routes/apiRoutes.js
@@ -2,24 +2,30 @@ var db = require("../models");
 
 module.exports = function(app) {
   // Get all examples
-  app.get("/api/examples", function(req, res) {
-    db.Example.findAll({}).then(function(dbExamples) {
+  app.get("/api/themes", function(req, res)
+  {
+    db.Example.findAll({})
+    .then(function(dbExamples) 
+    {
       res.json(dbExamples);
     });
   });
 
   // Create a new example
-  app.post("/api/examples", function(req, res) {
-    db.Example.create(req.body).then(function(dbExample) {
+  app.post("/api/examples", function(req, res) 
+  {
+    db.Example.create(req.body).then(function(dbExample) 
+    {
       res.json(dbExample);
     });
   });
 
   // Delete an example by id
-  app.delete("/api/examples/:id", function(req, res) {
-    db.Example.destroy({ where: { id: req.params.id } }).then(function(
-      dbExample
-    ) {
+  app.delete("/api/examples/:id", function(req, res) 
+  {
+    db.Example.destroy({ where: { id: req.params.id } })
+    .then(function(dbExample) 
+    {
       res.json(dbExample);
     });
   });
